requirements.txt
streamlit>=1.35
python-dotenv>=1.0.1
openai>=1.40.0
qdrant-client>=1.9.0
sqlmodel>=0.0.21
pydantic>=2.7
numpy>=1.26
requests>=2.31